+++ lib/src/log_file.bckp
@@ -4,10 +4,6 @@ import 'package:meta/meta.dart';
 import 'identifier.dart';
 import 'package:path/path.dart' as p;
 import 'dart:collection';
-import 'log_range.dart';
-import 'log_line.dart';
-import 'log_line_replay.dart';
-import 'log_line_serializer.dart';
 
 class LogFile {
   static final _newlineByte = utf8.encode(_newline).first;
@@ -25,8 +21,6 @@ class LogFile {
   void _openRaf() {
     _raf = File(path).openSync(mode: FileMode.append);
   }
-
-  LogLineSerializer lineSerializer(String key) => LogLineSerializer(_raf, key);
 
   LogRange appendLine(String data) {
     assert(
@@ -60,7 +54,7 @@ class LogFile {
   }
   List<int> _lastLineContainingBuffer() {
     var buffer = List<int>();
-    var chunksIterator = _BackwardsChunkIterable(_raf.lengthSync());
+    var chunksIterator = BackwardsChunkIterable(_raf.lengthSync());
     int foundNewlinesCount = 0;
     for (LogRange chunkRange in chunksIterator) {
       _raf.setPositionSync(chunkRange.start);
@@ -127,6 +121,15 @@ class LogFile {
   String toString() => File(path).readAsStringSync();
 }
 
+class LogRange {
+  final int start;
+  final int length;
+
+  LogRange(this.start, this.length);
+
+  String toString() => 'LogRange($start, $length)';
+}
+
 String _suffixedFilenamePath(String originalFilePath, String nameSuffix) {
   String name = p.basenameWithoutExtension(originalFilePath);
   String ext = p.extension(originalFilePath);
@@ -134,19 +137,76 @@ String _suffixedFilenamePath(String originalFilePath, String nameSuffix) {
   return p.join(base, '$name-$nameSuffix.$ext');
 }
 
-class _BackwardsChunkIterable extends IterableBase<LogRange> {
-  final int _length;
+class LogLine {
+  final LogRange range;
+  final String data;
 
-  _BackwardsChunkIterable(this._length);
-
-  Iterator<LogRange> get iterator => _BackwardsChunkIterator(_length);
+  LogLine(this.range, this.data);
 }
 
-class _BackwardsChunkIterator implements Iterator<LogRange> {
+class LogLineReplay extends IterableBase<LogLine> {
+  final String _path;
+  final int _newlineByte;
+
+  LogLineReplay(this._path, this._newlineByte);
+
+  Iterator<LogLine> get iterator => LogLineIterator(_path, _newlineByte);
+}
+
+class LogLineIterator implements Iterator<LogLine> {
+  final RandomAccessFile _raf;
+  final _buffer = List<int>();
+  final int _newlineByte;
+  int _start;
+  int _length;
+
+  LogLineIterator(String path, this._newlineByte)
+    : _raf = File(path).openSync(mode: FileMode.read)
+  {
+    _length = _raf.lengthSync();
+    _raf.setPositionSync(0);
+  }
+
+  bool moveNext() {
+    var hasNext = false;
+    _start = _raf.positionSync();
+    for (int i = _start; i < _length; i++) {
+      int byte = _raf.readByteSync();
+      if (byte == -1) break;
+      else if (byte == _newlineByte) {
+        hasNext = true;
+        break;
+      }
+      _buffer.add(byte);
+      hasNext = true;
+    }
+    if (!hasNext) {
+      _raf.closeSync();
+    }
+    return hasNext;
+  }
+
+  LogLine get current {
+    var range = LogRange(_start, _buffer.length);
+    var line = LogLine(range, utf8.decode(_buffer));
+    _buffer.clear();
+    return line;
+  }
+}
+
+class BackwardsChunkIterable extends IterableBase<LogRange> {
+  final int _length;
+
+  BackwardsChunkIterable(this._length);
+
+  Iterator<LogRange> get iterator => BackwardsChunkIterator(_length);
+}
+
+class BackwardsChunkIterator implements Iterator<LogRange> {
   final int _length;
   LogRange _current;
 
-  _BackwardsChunkIterator(this._length)
+  BackwardsChunkIterator(this._length)
     : _current = LogRange(_length, 0);
 
   bool moveNext() {
