+++ lib/src/log_line.dart
@@ -1,0 +1,8 @@
+import 'log_range.dart';
+
+class LogLine {
+  final LogRange range;
+  final String data;
+
+  LogLine(this.range, this.data);
+}
