+++ lib/src/log_range.dart
@@ -1,0 +1,8 @@
+class LogRange {
+  final int start;
+  final int length;
+
+  LogRange(this.start, this.length);
+
+  String toString() => 'LogRange($start, $length)';
+}
