+++ lib/src/todos.md
@@ -1,0 +1,2 @@
+1. Simplify serialization
+  - can be done directly into the file without circumvention via a serializer
